# Make validateToken a function